improved how users are imported
now we can just call the users/default.nix
to auto import all the users in the list
extraSpecialArgs.users :D

The only downside I can think of rn is maybe there'll be a system
where none of the users use home manager so we'll be importing the home
manager modules for no reason. But eh, silly edge case.